requirements.txt
numpy==1.25.2
matplotlib==3.7.2
pillow==10.1.0
